Added unique ID to Xray set
Added unique ID to count rows in X-ray set as a number of patient Ids were duplicated.  There was an issue when breaking this notebook into train / test / validation as a number of classes only occur a handful of times and it's very difficult to randomly split the dataset so each set contains exactly 11 classes.